+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes for Authorize.net for Craft Commerce
 
+## 1.0.6 - 2019-08-01
+
+{note} Significant changes have been made in this update, and a test transaction should be processed after upgrading this plugin. Saved payment sources are disabled by default, but can be enabled in the gateway settings.
+
+### Added
+- Support for Saved Payment Sources using the Authorize.net Customer Information Manager.
+
 ## 1.0.5 - 2019-07-23
 
 ### Changed